refactor: Move Repository to each domain location
