+++ src/main/java/com/psj/itembrowser/cart/repository/CartProductRelationRepository.java
@@ -1,4 +1,4 @@
-package com.psj.itembrowser.security.data.config;
+package com.psj.itembrowser.cart.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
