+++ src/main/java/com/psj/itembrowser/member/repository/MemberRepository.java
@@ -1,4 +1,4 @@
-package com.psj.itembrowser.security.data.config;
+package com.psj.itembrowser.member.repository;
 
 import java.util.Optional;
