+++ src/main/java/com/psj/itembrowser/order/repository/OrderRepository.java
@@ -1,4 +1,4 @@
-package com.psj.itembrowser.security.data.config;
+package com.psj.itembrowser.order.repository;
 
 import java.util.Optional;
