+++ src/main/java/com/psj/itembrowser/order/repository/OrdersProductRelationRepository.java
@@ -1,4 +1,4 @@
-package com.psj.itembrowser.security.data.config;
+package com.psj.itembrowser.order.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
