+++ src/main/java/com/psj/itembrowser/product/repository/ProductRepository.java
@@ -1,4 +1,4 @@
-package com.psj.itembrowser.security.data.config;
+package com.psj.itembrowser.product.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
